day09/index.test.ts
import { readFileSync } from "fs";
import { resolve } from "path";
import * as functions from ".";

describe("day 9", () => {
  describe("part 1", () => {
    test("sample data", () => {
      const data = readFileSync(resolve(__dirname, "sample.txt"), "utf-8");
      const result = functions.part1(data);
      expect(result).toBe(15);
    });

    test("actual data", () => {
      const data = readFileSync(resolve(__dirname, "actual.txt"), "utf-8");
      const result = functions.part1(data);
      expect(result).toBe(566);
    });
  });

  describe.only("part 2", () => {
    test("sample data", () => {
      const data = readFileSync(resolve(__dirname, "sample.txt"), "utf-8");
      const result = functions.part2(data);
      expect(result).toBe(1134);
    });

    test("actual data", () => {
      const data = readFileSync(resolve(__dirname, "actual.txt"), "utf-8");
      const result = functions.part2(data);
      expect(result).not.toBe(351900);
      expect(result).not.toBe(1017918);
      expect(result).not.toBe(948480);
      expect(result).toBe(891684);
    });
  });
});
